ref: Separate deckbrowser and overview code into separate files
