+++ tests/addon/test_integration.py
@@ -143,11 +143,6 @@ from ankihub.gui.config_dialog import (
     setup_config_dialog_manager,
 )
 from ankihub.gui.deck_updater import _AnkiHubDeckUpdater, ah_deck_updater
-from ankihub.gui.deckbrowser import (
-    FLASHCARD_SELECTOR_OPEN_BUTTON_ID,
-    FLASHCARD_SELECTOR_SYNC_NOTES_ACTIONS_PYCMD,
-    FlashCardSelectorDialog,
-)
 from ankihub.gui.decks_dialog import DeckManagementDialog
 from ankihub.gui.editor import SUGGESTION_BTN_ID
 from ankihub.gui.errors import upload_logs_and_data_in_background
@@ -169,6 +164,11 @@ from ankihub.gui.operations.new_deck_subscriptions import (
 )
 from ankihub.gui.operations.utils import future_with_result
 from ankihub.gui.optional_tag_suggestion_dialog import OptionalTagsSuggestionDialog
+from ankihub.gui.overview import (
+    FLASHCARD_SELECTOR_OPEN_BUTTON_ID,
+    FLASHCARD_SELECTOR_SYNC_NOTES_ACTIONS_PYCMD,
+    FlashCardSelectorDialog,
+)
 from ankihub.gui.reviewer import CLOSE_ANKIHUB_CHATBOT_PYCMD
 from ankihub.gui.suggestion_dialog import SuggestionDialog
 from ankihub.main.deck_creation import create_ankihub_deck, modify_note_type
